+++ custom_components/autoarm/const.py
@@ -276,3 +276,4 @@ class ChangeSource(StrEnum):
     SUNSET = auto()
     ZOMBIFICATION = auto()
     STARTUP = auto()
+    UNKNOWN = auto()